requirements.txt
streamlit
yfinance
pandas
numpy
plotly
requests
beautifulsoup4
fredapi
python-dotenv